Inherit standard faces where possible to make magit less hostile to color-theming
Just as org inherits various faces from outline-mode, there are plenty
of built-in faces which can serve as a logical base for magit's
faces. This patch adds :inherit to several faces, and there may be an
opportunity to add more. Failing to inherit faces where it is
reasonable to do so means that magit is likely to look odd for users
who have installed custom color-themes, which typically only redefine
the more basic faces.