Use nothing as an editor to make a false commit
When committing, and emacsclient is unavailable or unwanted, we still
want to run git commit to generate the initial commit message.

By deliberately using a GIT_EDITOR that do not exist, we make git try to
commit, but failing to do it. We can then edit COMMIT_EDITMSG to do the
real commit latter.